+++ src/main/java/net/commoble/jumbofurnace/client/JumboFurnaceScreen.java
@@ -2,6 +2,7 @@ package net.commoble.jumbofurnace.client;
 
 import net.commoble.jumbofurnace.JumboFurnace;
 import net.commoble.jumbofurnace.jumbo_furnace.JumboFurnaceMenu;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.network.chat.Component;
@@ -11,6 +12,8 @@ import net.minecraft.world.entity.player.Inventory;
 public class JumboFurnaceScreen extends AbstractContainerScreen<JumboFurnaceMenu>
 {
 	public static final ResourceLocation GUI_TEXTURE = JumboFurnace.id("textures/gui/jumbo_furnace.png");
+	public static final ResourceLocation LIT_PROGRESS_SPRITE = ResourceLocation.withDefaultNamespace("container/furnace/lit_progress");
+	private static final ResourceLocation COOK_PROGRESS_SPRITE = ResourceLocation.withDefaultNamespace("container/furnace/burn_progress");
 
 	// progress bar stuff
 	public static final int BURN_METER_FROM_X = 176;
@@ -20,8 +23,6 @@ public class JumboFurnaceScreen extends AbstractContainerScreen<JumboFurnaceMenu
 	public static final int BURN_METER_TO_X = 27;
 	public static final int BURN_METER_TO_Y = 73;
 	
-	public static final int COOK_METER_FROM_X = 176;
-	public static final int COOK_METER_FROM_Y = 14;
 	public static final int COOK_METER_WIDTH = 24;
 	public static final int COOK_METER_HEIGHT = 16;
 	public static final int COOK_METER_TO_X = 79;
@@ -59,10 +60,34 @@ public class JumboFurnaceScreen extends AbstractContainerScreen<JumboFurnaceMenu
 		if (this.menu.isBurning())
 		{
 			int burnAmount = (this.menu).getBurnLeftScaled() + 1;
-			graphics.blit(GUI_TEXTURE, xStart + BURN_METER_TO_X, yStart + BURN_METER_TO_Y + BURN_METER_HEIGHT - burnAmount, BURN_METER_FROM_X, BURN_METER_HEIGHT - burnAmount, BURN_METER_WIDTH, burnAmount);
+			graphics.blitSprite(LIT_PROGRESS_SPRITE, 14, 14, 0, 14 - burnAmount, xStart + BURN_METER_TO_X, yStart + BURN_METER_TO_Y + 14 - burnAmount, 14, burnAmount);
+//			graphics.blitSprite(LIT_PROGRESS_SPRITE, xStart + BURN_METER_TO_X, yStart + BURN_METER_TO_Y + BURN_METER_HEIGHT - burnAmount, BURN_METER_FROM_X, BURN_METER_HEIGHT - burnAmount, BURN_METER_WIDTH, burnAmount);
 		}
 
+		int cookMeterPixels = this.getCookMeterPixels(partialTicks);
+		if (cookMeterPixels > 0)
+		{
+			graphics.blitSprite(COOK_PROGRESS_SPRITE, COOK_METER_WIDTH, COOK_METER_HEIGHT, 0, 0, xStart + COOK_METER_TO_X, yStart + COOK_METER_TO_Y, cookMeterPixels, COOK_METER_HEIGHT);
+		}
 //		int cookProgress = (this.menu).getCookProgressionScaled() + 1;
 //		graphics.blit(GUI_TEXTURE, xStart + COOK_METER_TO_X, yStart + COOK_METER_TO_Y, COOK_METER_FROM_X, COOK_METER_FROM_Y, cookProgress, COOK_METER_HEIGHT);
+	}
+	
+	@SuppressWarnings("resource")
+	private int getCookMeterPixels(float partialTicks)
+	{
+		int recipes = this.menu.getCurrentRecipeCount();
+		if (recipes <= 0)
+		{
+			return 0;
+		}
+		if (this.menu.getBurnTimeRemaining() <= 0)
+		{
+			return COOK_METER_WIDTH / 2;
+		}
+		int extraRecipes = recipes - 1;
+		double gameTime = (Minecraft.getInstance().level.getGameTime() % COOK_METER_WIDTH) + partialTicks;
+		double scaledTime = gameTime * (extraRecipes/16); // 4x meter speed at max recipes
+		return (int)(scaledTime % COOK_METER_WIDTH);
 	}
 }